Separated planners and paths to their own folders
