+++ soccer/CMakeLists.txt
@@ -25,26 +25,27 @@ set(ROBOCUP_LIB_SRC
     "optimization/ParallelGradientAscent1D.cpp"
     "optimization/NelderMead2D.cpp"
     "optimization/PythonFunctionWrapper.cpp"
-    "planning/CompositePath.cpp"
+    "planning/paths/CompositePath.cpp"
     "planning/DynamicObstacle.cpp"
-    "planning/DirectTargetPathPlanner.cpp"
-    "planning/TuningPathPlanner.cpp"
-    "planning/EscapeObstaclesPathPlanner.cpp"
-    "planning/InterpolatedPath.cpp"
-    "planning/IndependentMultiRobotPathPlanner.cpp"
-    "planning/InterceptPlanner.cpp"
+    "planning/planners/DirectTargetPathPlanner.cpp"
+    "planning/planners/TuningPathPlanner.cpp"
+    "planning/planners/EscapeObstaclesPathPlanner.cpp"
+    "planning/paths/InterpolatedPath.cpp"
+    "planning/planners/IndependentMultiRobotPathPlanner.cpp"
+    "planning/planners/InterceptPlanner.cpp"
     "planning/MotionConstraints.cpp"
     "planning/RotationConstraints.cpp"
-    "planning/Path.cpp"
-    "planning/RRTPlanner.cpp"
-    "planning/RRTUtil.cpp"
-    "planning/PivotPathPlanner.cpp"
-    "planning/SettlePathPlanner.cpp"
-    "planning/CollectPathPlanner.cpp"
-    "planning/LineKickPlanner.cpp"
-    "planning/SingleRobotPathPlanner.cpp"
-    "planning/TargetVelPathPlanner.cpp"
-    "planning/TrapezoidalPath.cpp"
+    "planning/paths/Path.cpp"
+        "planning/planners/RRTPlanner.cpp"
+    "planning/planners/RRTUtil.cpp"
+    "planning/planners/PivotPathPlanner.cpp"
+    "planning/planners/SettlePathPlanner.cpp"
+    "planning/planners/CollectPathPlanner.cpp"
+    "planning/planners/LineKickPlanner.cpp"
+    "planning/planners/SingleRobotPathPlanner.cpp"
+    "planning/planners/TargetVelPathPlanner.cpp"
+    "planning/PlannerNode.cpp"
+    "planning/paths/TrapezoidalPath.cpp"
     "Processor.cpp"
     "radio/NetworkRadio.cpp"
     "radio/PacketConvert.cpp"
@@ -224,8 +225,8 @@ set(SOCCER_TEST_SRC
     "optimization/ParallelGradientAscent1DTest.cpp"
     "optimization/NelderMead2DTest.cpp"
     "planning/PathTest.cpp"
-    "planning/EscapeObstaclesPathPlannerTest.cpp"
-    "planning/TargetVelPathPlannerTest.cpp"
+    "planning/planners/EscapeObstaclesPathPlannerTest.cpp"
+    "planning/planners/TargetVelPathPlannerTest.cpp"
     "TestMain.cpp"
     "vision/tests/CameraBallTest.cpp"
     "vision/tests/KalmanBallTest.cpp"
