+++ soccer/planning/PlannerNode.cpp
@@ -1,0 +1,10 @@
+#include <planning/PlannerNode.hpp>
+
+namespace planning {
+PlannerNode::PlannerNode(Context* context) : context_{context} {}
+
+using namespace Geometry2d;
+void PlannerNode::run() {
+
+}
+}  // namespace planning
