+++ soccer/planning/planners/DirectTargetPathPlanner.cpp
@@ -1,27 +1,34 @@
-#include "TuningPathPlanner.hpp"
-#include "MotionCommand.hpp"
+#include "DirectTargetPathPlanner.hpp"
 
+#include "planning/MotionCommand.hpp"
+
+using namespace Geometry2d;
 namespace Planning {
 
-std::unique_ptr<Path> TuningPathPlanner::run(PlanRequest& planRequest) {
+double vectorInDirection(Point point, Point direction) {
+    const auto vector = point.dot(direction.normalized());
+    return std::max(vector, 0.0);
+}
+
+std::unique_ptr<Path> DirectTargetPathPlanner::run(PlanRequest& planRequest) {
     const MotionInstant& startInstant = planRequest.start;
     const auto& motionConstraints = planRequest.constraints.mot;
     const Geometry2d::ShapeSet& obstacles = planRequest.obstacles;
     std::unique_ptr<Path>& prevPath = planRequest.prevPath;
 
-    const Planning::TuningPathCommand& command =
-        dynamic_cast<const Planning::TuningPathCommand&>(
+    const Planning::DirectPathTargetCommand& command =
+        dynamic_cast<const Planning::DirectPathTargetCommand&>(
             *planRequest.motionCommand);
 
     if (shouldReplan(planRequest)) {
         Geometry2d::Point endTarget = command.pathGoal.pos;
+        const auto direction = (endTarget - startInstant.pos).normalized();
+
         float endSpeed = command.pathGoal.vel.mag();
-        // Tells the robot that is actually in a different location
-        // This forces the PID Tuner to kick in to move the robot to its new
-        // location
-        auto path = std::unique_ptr<Path>(
-            new TrapezoidalPath(endTarget, startInstant.vel.mag(), endTarget,
-                                endSpeed, motionConstraints));
+        auto path = std::make_unique<TrapezoidalPath>(
+            startInstant.pos, vectorInDirection(startInstant.vel, direction),
+            endTarget, vectorInDirection(command.pathGoal.vel, direction),
+            motionConstraints);
         path->setStartTime(RJ::now());
         return std::move(path);
     } else {
@@ -29,18 +36,21 @@ std::unique_ptr<Path> TuningPathPlanner::run(PlanRequest& planRequest) {
     }
 }
 
-bool TuningPathPlanner::shouldReplan(const PlanRequest& planRequest) const {
+bool DirectTargetPathPlanner::shouldReplan(
+    const PlanRequest& planRequest) const {
     const MotionConstraints& motionConstraints = planRequest.constraints.mot;
     const Geometry2d::ShapeSet& obstacles = planRequest.obstacles;
     const Path* prevPath = planRequest.prevPath.get();
 
-    const Planning::TuningPathCommand& command =
-        dynamic_cast<const Planning::TuningPathCommand&>(
+    const Planning::DirectPathTargetCommand& command =
+        dynamic_cast<const Planning::DirectPathTargetCommand&>(
             *planRequest.motionCommand);
 
     if (!prevPath) {
         return true;
     } else {
+        // For DirectTarget commands, we replan if the goal position or velocity
+        // have changed beyond a certain threshold
         Geometry2d::Point endTarget = command.pathGoal.pos;
         float endSpeed = command.pathGoal.vel.mag();
         float targetPosChange = (prevPath->end().motion.pos - endTarget).mag();
@@ -49,7 +59,7 @@ bool TuningPathPlanner::shouldReplan(const PlanRequest& planRequest) const {
         if (targetPosChange >
                 SingleRobotPathPlanner::goalPosChangeThreshold() ||
             targetVelChange >
-                SingleRobotPathPlanner::goalVelChangeThreshold()) {
+                SingleRobotPathPlanner::goalPosChangeThreshold()) {
             // FIXME: goalChangeThreshold shouldn't be used for checking
             // speed differences as it is in the above 'if' statement
             return true;
