+++ soccer/planning/planners/SettlePathPlanner.cpp
@@ -3,10 +3,10 @@
 #include <algorithm>
 #include <cmath>
 
-#include "CompositePath.hpp"
 #include "Configuration.hpp"
 #include "Constants.hpp"
-#include "MotionInstant.hpp"
+#include "planning/MotionInstant.hpp"
+#include "planning/paths/CompositePath.hpp"
 
 using namespace std;
 using namespace Geometry2d;
